Move doctests to examples
