+++ exonum/Cargo.toml
@@ -66,4 +66,3 @@ float_serialize = []
 long_benchmarks = []
 flame_profile = ["exonum_profiler/nomock", "exonum_flamer"]
 metrics-log = []
-doctests = []
